pyproject.toml
[tool.poetry]
name = "business-analyst-ai-agent"
version = "0.1.0"
description = "Business Analyst AI agent skeleton with FastAPI, LangChain, Gemini, Neo4j, Pinecone, Google Docs"
authors = ["Auto Generated"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
langchain = "^0.0.300"
neo4j = "^5.11.0"
pinecone-client = "^1.1.0"
google-api-python-client = "^2.88.0"
google-auth = "^2.20.0"
httpx = "^0.24.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^24.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry_core.masonry.api"
